refactor: Member 필드명 변경(@JvmField 적용), 하드코딩 문자열 제거, JWT Claims 불필요한 필드 정리
