+++ prestudy/README.md
@@ -2,8 +2,9 @@
 - 게시판 만들기
 
 ### Architecture
-- 클린 아키텍처 적용
+- 기본적으로 클린 아키텍처 적용
 - 모듈 설명
+  - `auth` : 인증/인가 관련 모듈, Spring Security 의존성이 높아 레이어드 아키텍처 적용
   - `bootstrap` : 스프링 애플리케이션 구동 모듈
   - `common` : 도메인 모듈에 공통적으로 사용하는 클래스를 모은 모듈
   - `domain` : 각 도메인에 해당하는 모듈
